move of convertTypes, test update
