+++ src/rdb/lib/convertTypes.cc
@@ -1,4 +1,4 @@
-#include "convertTypes.h"
+#include "rdb/convertTypes.h"
 
 #include <spdlog/spdlog.h>
