+++ src/retractor/parser.cpp
@@ -7,12 +7,14 @@
 #include <iostream>
 #include <string>
 
+#include <rdb/convertTypes.h>
+
 #include "Parser/RQLBaseListener.h"
 #include "Parser/RQLLexer.h"
 #include "Parser/RQLParser.h"
 #include "QStruct.h"
 #include "antlr4-runtime/antlr4-runtime.h"
-#include "convertTypes.h"
+
 
 extern qTree coreInstance;
